requirements.txt
protobuf==3.20.0
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.9
django-debug-toolbar==3.2.4
django-heroku==0.3.1
psycopg2==2.9.2
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
tensorflow-cpu==2.9.0
protobuf==3.20.0